Remove CEP rule
It'd depend on whether the business deems cep validation important
enough to be worth validating remotely. When creating a single
patient it should be fine, since the frontend would probably hit
our CEP endpoint to autocomplete the address and caching it as a
result anyway. But it'd probably be an issue on imports.